@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import {} from '@angular/common/http';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { RouterModule, RouterOutlet } from '@angular/router';
 import { Solution1Component } from "./Components/solution1/solution1.component";
@@ -40,7 +40,11 @@ import { Solution9Component } from "./Components/solution9/solution9.component";
       <router-outlet></router-outlet>
     `,
     imports: [
-        RouterModule, RouterOutlet, HttpClientModule, CommonModule,
+        RouterModule, RouterOutlet, 
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule, CommonModule,
         Solution1Component,
         Solution2Component,
         Solution3Component,
